+++ crates/cli/src/app.rs
@@ -27,7 +27,7 @@ pub enum Commands {
         long_about = "Add an alias to a tool, that maps to a specific version, or another alias."
     )]
     Alias {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
 
         #[arg(required = true, help = "Alias name")]
@@ -43,7 +43,7 @@ pub enum Commands {
         long_about = "Display the absolute path to a tools binary. If no version is provided,\nit will detected from the current environment."
     )]
     Bin {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
 
         #[arg(help = "Version of tool")]
@@ -81,7 +81,7 @@ pub enum Commands {
         long_about = "Download and install a tool by unpacking the archive to ~/.proto/tools."
     )]
     Install {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
 
         #[arg(default_value = "latest", help = "Version of tool")]
@@ -102,7 +102,7 @@ pub enum Commands {
         long_about = "Install a global dependency for the specified tool. Depending on the tool, the dependency will either be installed to ~/.proto/tools/<tool>/globals or ~/<tool>."
     )]
     InstallGlobal {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
 
         #[arg(required = true, help = "Dependencies and optional version to install")]
@@ -115,7 +115,7 @@ pub enum Commands {
         long_about = "Set the global default version of a tool. This will pin the version in the ~/.proto/tools installation directory."
     )]
     Global {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
 
         #[arg(required = true, help = "Version of tool")]
@@ -129,7 +129,7 @@ pub enum Commands {
         long_about = "List installed versions by scanning the ~/.proto/tools directory for possible versions."
     )]
     List {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
     },
 
@@ -140,7 +140,7 @@ pub enum Commands {
         long_about = "List installed globals by scanning the global bins installation directory. Will return the canonical source path."
     )]
     ListGlobal {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
     },
 
@@ -151,7 +151,7 @@ pub enum Commands {
         long_about = "List available versions by resolving versions from the tool's remote release manifest."
     )]
     ListRemote {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
     },
 
@@ -161,7 +161,7 @@ pub enum Commands {
         long_about = "Set the local version of a tool. This will create a .prototools file (if it does not exist)\nin the current working directory with the appropriate tool and version."
     )]
     Local {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
 
         #[arg(required = true, help = "Version of tool")]
@@ -175,7 +175,7 @@ pub enum Commands {
         long_about = "Run a tool after detecting a version from the environment. In order of priority,\na version will be resolved from a provided CLI argument, a PROTO_VERSION environment variable,\na local version file (.prototools), and lastly a global version file (~/.proto/tools/version).\n\nIf no version can be found, the program will exit with an error."
     )]
     Run {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
 
         #[arg(help = "Version of tool")]
@@ -200,7 +200,7 @@ pub enum Commands {
 
     #[command(name = "ua", name = "unalias", about = "Remove an alias from a tool.")]
     Unalias {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
 
         #[arg(required = true, help = "Alias name")]
@@ -214,7 +214,7 @@ pub enum Commands {
         long_about = "Uninstall a tool and remove the installation from ~/.proto/tools."
     )]
     Uninstall {
-        #[arg(required = true, value_enum, help = "Type of tool")]
+        #[arg(required = true, help = "Type of tool")]
         tool: ToolType,
 
         #[arg(required = true, help = "Version of tool")]