+++ crates/cli/src/commands/bin.rs
@@ -7,7 +7,7 @@ pub async fn bin(
     forced_version: Option<String>,
     use_shim: bool,
 ) -> SystemResult {
-    let mut tool = create_tool(&tool_type)?;
+    let mut tool = create_tool(&tool_type).await?;
     let manifest = Manifest::load(tool.get_manifest_path())?;
     let version = detect_version(&tool, &manifest, forced_version).await?;