+++ crates/schema-plugin/src/schema.rs
@@ -72,10 +72,12 @@ impl Default for ShimSchema {
 }
 
 #[derive(Debug, Default, Deserialize)]
-pub enum ToolType {
+#[serde(rename_all = "kebab-case")]
+pub enum SchemaToolType {
     #[default]
     Language,
     DependencyManager,
+    Cli,
 }
 
 #[derive(Debug, Default, Deserialize)]
@@ -84,7 +86,7 @@ pub struct Schema {
     pub bin: String,
     pub name: String,
     #[serde(rename = "type")]
-    pub type_of: ToolType,
+    pub type_of: SchemaToolType,
     pub platform: FxHashMap<String, PlatformMapper>,
 
     pub detect: DetectSchema,