+++ crates/schema-plugin/tests/schema_test.rs
@@ -16,7 +16,7 @@ fn create_plugin(dir: &Path, mut schema: Schema) -> SchemaPlugin {
     schema.name = "moon".into();
     schema.bin = "moon".into();
 
-    let mut tool = SchemaPlugin::new(Proto::from(dir), schema);
+    let mut tool = SchemaPlugin::new(Proto::from(dir), "moon".into(), schema);
     tool.version = Some("1.0.0".into());
     tool
 }
@@ -131,6 +131,7 @@ mod schema_plugin {
                 Proto::from(fixture.path())
                     .temp_dir
                     .join("moon")
+                    .join("1.0.0")
                     .join(tool.get_download_file().unwrap())
             );
 
@@ -549,6 +550,7 @@ mod schema_plugin {
                 Proto::from(fixture.path())
                     .temp_dir
                     .join("moon")
+                    .join("1.0.0")
                     .join(tool.get_checksum_file().unwrap())
             );